requirements.txt
pandas
google-generativeai
chromadb
gspread
oauth2client
requests
python-dotenv
flask
flask-cors
mlflow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
azure-storage-blob